Reorganize session utils
