+++ shiny/session/__init__.py
@@ -3,5 +3,12 @@ Tools for working within a (user) session context.
 """
 
 from ._session import *
+from ._utils import *
 
-__all__ = ("Session", "Inputs", "Outputs", "get_current_session")
+__all__ = (
+    "Session",
+    "Inputs",
+    "Outputs",
+    "get_current_session",
+    "require_active_session",
+)
