+++ shiny/ui/_progress.py
@@ -3,8 +3,7 @@ __all__ = ("Progress",)
 from typing import Optional, Dict, Any
 from warnings import warn
 from .._utils import run_coro_sync, rand_hex
-from ..session import Session
-from ..session._utils import require_active_session
+from ..session import Session, require_active_session
 
 
 class Progress:
